Change T_DIR, T_FILE, and T_DEVICE to i16

+++ kernel-rs/src/fs/path.rs
@@ -143,7 +143,7 @@ impl Path {
             path = new_path;
 
             (*ip).lock();
-            if (*ip).typ as i32 != T_DIR {
+            if (*ip).typ != T_DIR {
                 (*ip).unlockput();
                 return Err(());
             }
